Split thin rooms properly
Split horizontal and split vertical now return a bool for whether the split worked or not. Changed the split unction to try splitting the room the other way if the first attempt doesn't work. changed display to show the room 1 bigger than it is, this makes the rooms look like they have wall in between. If there are no more rooms to split the loop stops splitting rooms.