Added coroutine funtionality
Changed the start function into a coroutine. I took out some old code. I added a return bool to the split room funtion, so if the room doesn't split the coroutine doesn't wait. Reworked the color code, so the color is generated when the room is generated instead of at the end. This is omportant for visualizing the work.