Reduced nesting and reworked spliting code
Changed a nesting if statement to a conditional break then the next if statement. Improves readablility. Also changed the math for the splitting. Old system calculates rooms using sizes and positions. New system calculates rooms using min and max positions. This is easier to read and change. Enabled coroutine for debugging.