improved readability and memory usage
changed hashset from storing worldspace positions to storing gridspace positions. conversion to world space is done in one place right before instanciation. this saves memory, as the y coordinate is nolonger stored in the hash set, which was only 0s anyway. also reduced code duplication by converting to worldspace in one place. improved variable names. storing the offset for walls and floors in a variable instead of hardcoding it.